clients/settings: add /promote and /settings
