+++ clients/apps/web/src/pages/settings/index.tsx
@@ -1,21 +1,34 @@
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
-import LoadingScreen from '@/components/Dashboard/LoadingScreen'
-import Layout from '@/components/Layout/EmptyLayout'
-import type { NextPageWithLayout } from '@/utils/next'
-import { useRouter } from 'next/router'
-import { ReactElement, useEffect } from 'react'
+import DashboardLayout from '@/components/Layout/DashboardLayout'
+import NotificationSettings from '@/components/Settings/NotificationSettings'
+import { NextLayoutComponentType } from 'next'
+import Head from 'next/head'
+import { ReactElement } from 'react'
 
-const Page: NextPageWithLayout = () => {
-  const router = useRouter()
-  useEffect(() => {
-    router.push(`/settings/personal`)
-  })
-
+const Page: NextLayoutComponentType = () => {
   return (
     <>
-      <LoadingScreen>
-        <>Redirecting...</>
-      </LoadingScreen>
+      <Head>
+        <title>Polar | Settings</title>
+      </Head>
+
+      <DashboardLayout>
+        <>
+          <div className="relative z-0">
+            <div className="divide-y divide-gray-200 dark:divide-gray-800">
+              <Section>
+                <>
+                  <SectionDescription
+                    title="Email notifications"
+                    description="Polar will send emails for the notifications enabled below."
+                  />
+                  <NotificationSettings />
+                </>
+              </Section>
+            </div>
+          </div>
+        </>
+      </DashboardLayout>
     </>
   )
 }
@@ -23,8 +36,31 @@ const Page: NextPageWithLayout = () => {
 Page.getLayout = (page: ReactElement) => {
   return (
     <Gatekeeper>
-      <Layout>{page}</Layout>
+      <>{page}</>
     </Gatekeeper>
+  )
+}
+
+const Section = ({ children }: { children: ReactElement }) => {
+  return (
+    <div className="mb-4 flex flex-col space-y-4 pt-4 xl:flex-row-reverse xl:space-y-0">
+      {children}
+    </div>
+  )
+}
+
+const SectionDescription = ({
+  title,
+  description,
+}: {
+  title: string
+  description: string
+}) => {
+  return (
+    <div className="flex-shrink-0 xl:ml-8 xl:w-60">
+      <h2 className="mb-2 font-medium">{title}</h2>
+      <p className="text-sm text-gray-500 dark:text-gray-400">{description}</p>
+    </div>
   )
 }
